requirements.txt
asgiref==3.4.1
Django==3.0.3
django-rest-auth==0.9.5
djangorestframework==3.13.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
